agentic-chatbot/requirements.txt
pymongo==4.6.0
google-generativeai==0.3.2
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0